+++ src/test/java/domain/menu/MenuRepositoryTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.menu;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
