+++ src/test/java/domain/table/TableRepositoryTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
